Specifying analytical decisions in a specification setup

Creates all possible specifications as a combination of different dependent and independent variables, model types, control variables, potential subset analyses, as well as potentially other analytic choices. This function represents the first step in the analytic framework implemented in the package specr. The resulting class specr.setup then needs to be passed to the core function of the package called specr(), which fits the specified models across all specifications.

setup(
  data,
  x,
  y,
  model,
  ...,
  controls = NULL,
  add_to_formula = NULL,
  fun1 = function(x) broom::tidy(x, conf.int = TRUE),
  fun2 = function(x) broom::glance(x),
  simplify = FALSE
)

Arguments

data

The data set that should be used for the analysis

x

A vector denoting independent variables

y

A vector denoting the dependent variables

model

A vector denoting the model(s) that should be estimated.

...

Specification of potential subsets/groups. As the subsetting variable should be in the data set, please specify as follows (distinct(data, variable)). Note: This variable needs to be a numeric (<dbl> or <int>) or a character (<chr>) variable in the the data set! It will not work with a factor (<fct>) variable. If your grouping variable is a factor, please recode to a character variable beforehand. See also examples further below.

controls

A vector of the control variables that should be included. Defaults to NULL.

add_to_formula

A string specifying aspects that should always be included in the formula (e.g. a constant covariate, random effect structures...)

fun1

A function that extracts the parameters of interest from the fitted models. Defaults to tidy, which works with a large range of different models.

fun2

A function that extracts fit indices of interest from the models. Defaults to glance, which works with a large range of different models. Note: Different models result in different fit indices. Thus, if you use different models within one specification curve analysis, this may not work. In this case, you can simply set fun2 = NULL to not extract any fit indices.

simplify

Logical value indicating what type of combinations between control variables should be included in the specification. If FALSE (default), all combinations between the provided variables are created (none, each individually, each combination between each variable, all variables). If TRUE, only no covariates, each individually, and all covariates are included as specifications (akin to the default in specr version 0.2.1).

Value

An object of class specr.setup which includes all possible specifications based on combinations of the analytic choices. The resulting list includes a specification tibble, the data set, and additional information about the universe of specifications. Use methods(class = "specr.setup") for an overview on available methods.

Details

Empirical results are often contingent on analytical decisions that are equally defensible, often arbitrary, and motivated by different reasons. This decisions may introduce bias or at least variability. To this end, specification curve analyses (Simonsohn et al., 2020) or multiverse analyses (Steegen et al., 2016) refer to identifying the set of theoretically justified, statistically valid (and potentially also non-redundant specifications, fitting the "multiverse" of models represented by these specifications and extract relevant parameters often to display the results graphically as a so-called specification curve. This allows readers to identify consequential specifications decisions and how they affect the results or parameter of interest.

Use of this function

A general overview is provided in the vignettes vignette("specr"). It is assumed that you want to estimate the relationship between two variables (x and y). What varies may be what variables should be used for x and y, what model should be used to estimate the relationship, whether the relationship should be estimated for certain subsets, and whether different combinations of control variables should be included. This allows to (re)produce almost any analytical decision imaginable. See examples below for how a number of typical analytical decision can be implemented. Afterwards you pass the resulting object of a class specr.setup to the function specr() to run the specification curve analysis.

Note, the resulting class of specr.setup allows to use generic functions. Use methods(class = "specr.setup") for an overview on available methods and e.g., ?summary.specr.setup to view the dedicated help page.

References

Simonsohn, U., Simmons, J.P. & Nelson, L.D. (2020). Specification curve analysis. Nature Human Behaviour, 4, 1208–1214. https://doi.org/10.1038/s41562-020-0912-z

Steegen, S., Tuerlinckx, F., Gelman, A., & Vanpaemel, W. (2016). Increasing Transparency Through a Multiverse Analysis. Perspectives on Psychological Science, 11(5), 702-712. https://doi.org/10.1177/1745691616658637

See also

Examples

## Example 1 ----
# Setting up typical specifications
specs <- setup(data = example_data,             # data to be used
               x = c("x1", "x2"),               # independent variables
               y = c("y1", "y2"),               # dependent variables
               model = "lm",                    # model estimation function
               distinct(example_data, group1),  # first grouping variable for subsets
               distinct(example_data, group2),  # second grouping variable for subsets
               controls = c("c1", "c2", "c3"),  # control variables
               simplify = TRUE)                 # limited combinations of controls

# Check specifications
summary(specs, rows = 18)
#> Setup for the Specification Curve Analysis
#> -------------------------------------------
#> Class:                      specr.setup -- version: 1.0.0 
#> Number of specifications:   240 
#> 
#> Specifications:
#> 
#>   Independent variable:     x1, x2 
#>   Dependent variable:       y1, y2 
#>   Models:                   lm 
#>   Covariates:               no covariates, c1, c2, c3, all covariates 
#>   Subsets analyses:         middle & female, old & female, young & female, female, middle & male, old & male, young & male, male, middle, old, young, all 
#> 
#> Function used to extract parameters:
#> 
#>   function (x) 
#> broom::tidy(x, conf.int = TRUE)
#> <environment: 0x7f821a93beb0>
#> 
#> 
#> Head of specifications table (first 18 rows):
#> 
#> # A tibble: 18 × 8
#>    x     y     model controls      subsets         group1 group2 formula     
#>    <chr> <chr> <chr> <chr>         <chr>           <fct>  <fct>  <glue>      
#>  1 x1    y1    lm    no covariates middle & female middle female y1 ~ x1 + 1 
#>  2 x1    y1    lm    no covariates old & female    old    female y1 ~ x1 + 1 
#>  3 x1    y1    lm    no covariates young & female  young  female y1 ~ x1 + 1 
#>  4 x1    y1    lm    no covariates female          NA     female y1 ~ x1 + 1 
#>  5 x1    y1    lm    no covariates middle & male   middle male   y1 ~ x1 + 1 
#>  6 x1    y1    lm    no covariates old & male      old    male   y1 ~ x1 + 1 
#>  7 x1    y1    lm    no covariates young & male    young  male   y1 ~ x1 + 1 
#>  8 x1    y1    lm    no covariates male            NA     male   y1 ~ x1 + 1 
#>  9 x1    y1    lm    no covariates middle          middle NA     y1 ~ x1 + 1 
#> 10 x1    y1    lm    no covariates old             old    NA     y1 ~ x1 + 1 
#> 11 x1    y1    lm    no covariates young           young  NA     y1 ~ x1 + 1 
#> 12 x1    y1    lm    no covariates all             NA     NA     y1 ~ x1 + 1 
#> 13 x1    y1    lm    c1            middle & female middle female y1 ~ x1 + c1
#> 14 x1    y1    lm    c1            old & female    old    female y1 ~ x1 + c1
#> 15 x1    y1    lm    c1            young & female  young  female y1 ~ x1 + c1
#> 16 x1    y1    lm    c1            female          NA     female y1 ~ x1 + c1
#> 17 x1    y1    lm    c1            middle & male   middle male   y1 ~ x1 + c1
#> 18 x1    y1    lm    c1            old & male      old    male   y1 ~ x1 + c1

## Example 2 ----
# Setting up specifications for multilevel models
specs <- setup(data = example_data,
               x = c("x1", "x2"),
               y = c("y1", "y2"),
               model = c("lmer"),               # multilevel model
               distinct(example_data, group1),
               controls = c("c1", "c2"),
               add_to_formula = "(1|group2)")   # random effect in all models

# Check specifications
summary(specs)
#> Setup for the Specification Curve Analysis
#> -------------------------------------------
#> Class:                      specr.setup -- version: 1.0.0 
#> Number of specifications:   64 
#> 
#> Specifications:
#> 
#>   Independent variable:     x1, x2 
#>   Dependent variable:       y1, y2 
#>   Models:                   lmer 
#>   Covariates:               no covariates, c1, c2, c1 + c2 
#>   Subsets analyses:         middle, old, young, all 
#> 
#> Function used to extract parameters:
#> 
#>   function (x) 
#> broom::tidy(x, conf.int = TRUE)
#> <environment: 0x7f824091fc20>
#> 
#> 
#> Head of specifications table (first 6 rows):
#> 
#> # A tibble: 6 × 7
#>   x     y     model controls      subsets group1 formula                  
#>   <chr> <chr> <chr> <chr>         <chr>   <fct>  <glue>                   
#> 1 x1    y1    lmer  no covariates middle  middle y1 ~ x1 + 1 + (1|group2) 
#> 2 x1    y1    lmer  no covariates old     old    y1 ~ x1 + 1 + (1|group2) 
#> 3 x1    y1    lmer  no covariates young   young  y1 ~ x1 + 1 + (1|group2) 
#> 4 x1    y1    lmer  no covariates all     NA     y1 ~ x1 + 1 + (1|group2) 
#> 5 x1    y1    lmer  c1            middle  middle y1 ~ x1 + c1 + (1|group2)
#> 6 x1    y1    lmer  c1            old     old    y1 ~ x1 + c1 + (1|group2)

## Example 3 ----
# Setting up specifications with a different parameter extract functions

# Create custom extract function to extract different parameter and model
tidy_99 <- function(x) {
  fit <- broom::tidy(x,
     conf.int = TRUE,
     conf.level = .99)         # different alpha error rate
  fit$full_model = list(x)     # include entire model fit object as list
  return(fit)
}

# Setup specs
specs <- setup(data = example_data,
               x = c("x1", "x2"),
               y = c("y1", "y2"),
               model = "lm",
               fun1 = tidy_99,             # pass new function to setup
               add_to_formula = "c1 + c2") # set of covariates in all models

# Check specifications
summary(specs)
#> Setup for the Specification Curve Analysis
#> -------------------------------------------
#> Class:                      specr.setup -- version: 1.0.0 
#> Number of specifications:   4 
#> 
#> Specifications:
#> 
#>   Independent variable:     x1, x2 
#>   Dependent variable:       y1, y2 
#>   Models:                   lm 
#>   Covariates:               no covariates 
#>   Subsets analyses:         all 
#> 
#> Function used to extract parameters:
#> 
#>   function(x) {
#>   fit <- broom::tidy(x,
#>      conf.int = TRUE,
#>      conf.level = .99)         # different alpha error rate
#>   fit$full_model = list(x)     # include entire model fit object as list
#>   return(fit)
#> }
#> <environment: 0x7f821a88f548>
#> 
#> 
#> Head of specifications table (first 6 rows):
#> 
#> # A tibble: 4 × 6
#>   x     y     model controls      subsets formula              
#>   <chr> <chr> <chr> <chr>         <chr>   <glue>               
#> 1 x1    y1    lm    no covariates all     y1 ~ x1 + 1 + c1 + c2
#> 2 x1    y2    lm    no covariates all     y2 ~ x1 + 1 + c1 + c2
#> 3 x2    y1    lm    no covariates all     y1 ~ x2 + 1 + c1 + c2
#> 4 x2    y2    lm    no covariates all     y2 ~ x2 + 1 + c1 + c2