Fit models across all specifications

Runs the specification/multiverse analysis across specified models. This is the central function of the package and represent the second step in the analytic framework implemented in the package specr. It estimates and returns respective parameters and estimates of models that were specified via setup().

specr(x, data = NULL, workers = availableCores(), ..., message = TRUE)

Arguments

x

A specr.setup object resulting from setup or a tibble that contains the relevant specifications (e.g., a tibble resulting from as_tibble(setup(...))).

data

If x is not an object of "specr.setup" and simply a tibble, you need to provide the data set that should be used. Defaults to NULL as it is assumend that most users will create an object of class "specr.setup" that they'll pass to specr().

workers

A numeric value that indicates how many cores should be used. Defaults to availableCores() from the future package and thus uses all available cores. Note: If the models specified via setup() are not very computationally intensive, the overhead introduced trhough the parallelization (setting up several workers and R sessions, etc.) may actually make the fitting process slower. It only becomes considerably faster if the model fitting process is so intensive that the overhead no longer matters. In other words, in a lot of cases, setting workers = 1 can be a meaningful choice.

...

Further arguments that can be passed to future_pmap. This only becomes important if parallelization is used (workers > 1) and a custom model function is used. See examples 2 and 3 below for further details.

message

Logical value; indicates whether a short message should be printed after the estimation including information about the elapsed time. Defaults to TRUE.

Value

An object of class specr.object, which includes a data frame with all specifications their respective results along with many other useful information about the model. Parameters are extracted via the function passed to setup. By default this is broom::tidy() and the function broom::glance()).Several other aspects and information are included in the resulting class (e.g., number of specifications, time elapsed, subsets included in the analyses). Use methods(class = "specr.object") for an overview on available methods.

Details

Empirical results are often contingent on analytical decisions that are equally defensible, often arbitrary, and motivated by different reasons. This decisions may introduce bias or at least variability. To this end, specification curve analyses (Simonsohn et al., 2019) or multiverse analyses (Steegen et al., 2016) refer to identifying the set of theoretically justified, statistically valid (and potentially also non-redundant specifications, fitting the "multiverse" of models represented by these specifications and estract relevant parameters often to display the results graphically as a so-called specification curve. This allows readers to identify consequential specifications decisions and how they affect the results or parameter of interest.

Use of this function

A general overview is provided in the vignettes vignette("specr"). Generally, you create relevant specification using the function setup(). You then pass the resulting object of a class specr.setup to the present function specr() to run the specification curve analysis. Further note that the resulting object of class specr.object allows to use several generic function such as summary() or plot(). Use methods(class = "specr.object") for an overview on available methods and e.g., ?plot.specr.object to view the dedicated help page.

Disclaimer

We do see a lot of value in investigating how analytical choices affect a statistical outcome of interest. However, we strongly caution against using specr as a tool to somehow arrive at a better estimate compared to a single model. Running a specification curve analysis does not make your findings any more reliable, valid or generalizable than a single analysis. The method is meant to inform about the effects of analytical choices on results, and not a better way to estimate a correlation or effect.

References

Simonsohn, U., Simmons, J. P., & Nelson, L. D. (2019). Specification Curve: Descriptive and Inferential Statistics for all Plausible Specifications. Available at: https://doi.org/10.2139/ssrn.2694998

Steegen, S., Tuerlinckx, F., Gelman, A., & Vanpaemel, W. (2016). Increasing Transparency Through a Multiverse Analysis. Perspectives on Psychological Science, 11(5), 702-712. https://doi.org/10.1177/1745691616658637

See also

Examples

## Example 1 ----
# Setup up typical specifications
specs <- setup(data = example_data,            # providing data
               y = c("y1", "y2"),              # different y variables
               x = c("x1", "x2"),              # different x variables
               model = "lm",                   # only one model type
               distinct(example_data, group1), # subsets
               controls = c("c1", "c2"))       # Control these variables
# Run analysis (not parallelized, using simple `map` functions)
results <- specr(specs, workers = 1)
#> Models fitted based on 64 specifications
#> Cores used: 1 
#> 0.654 sec elapsed

# Summary of the results
summary(results)
#> Results of the specification curve analysis
#> -------------------
#> Technical details:
#> 
#>   Class:                          specr.object -- version: 0.3.0 
#>   Cores used:                     1 
#>   Duration of fitting process:    0.654 sec elapsed 
#>   Number of specifications:       64 
#> 
#> Descriptive summary of the specification curve:
#> 
#>  median  mad   min  max   q25  q75
#>    0.13 0.41 -0.39 0.66 -0.14 0.39
#> 
#> Descriptive summary of sample sizes: 
#> 
#>  median min  max
#>   338.5 323 1000
#> 
#> Head of the specification results (first 6 rows): 
#> 
#> # A tibble: 6 × 25
#>   x     y     model controls subsets group1 formula estimate std.error statistic
#>   <chr> <chr> <chr> <chr>    <chr>   <fct>  <glue>     <dbl>     <dbl>     <dbl>
#> 1 x1    y1    lm    1        middle  middle y1 ~ x…     0.61      0.07      9.28
#> 2 x1    y1    lm    1        old     old    y1 ~ x…     0.66      0.06     10.4 
#> 3 x1    y1    lm    1        young   young  y1 ~ x…     0.57      0.06     10.2 
#> 4 x1    y1    lm    1        all     NA     y1 ~ x…     0.62      0.04     16.4 
#> 5 x1    y1    lm    c1       middle  middle y1 ~ x…     0.6       0.07      8.81
#> 6 x1    y1    lm    c1       old     old    y1 ~ x…     0.64      0.06      9.81
#> # … with 15 more variables: p.value <dbl>, conf.low <dbl>, conf.high <dbl>,
#> #   fit_r.squared <dbl>, fit_adj.r.squared <dbl>, fit_sigma <dbl>,
#> #   fit_statistic <dbl>, fit_p.value <dbl>, fit_df <dbl>, fit_logLik <dbl>,
#> #   fit_AIC <dbl>, fit_BIC <dbl>, fit_deviance <dbl>, fit_df.residual <dbl>,
#> #   fit_nobs <dbl>
summary(results, type = "curve", subsets)
#> # A tibble: 4 × 8
#>   subsets median   mad    min   max    q25   q75   obs
#>   <chr>    <dbl> <dbl>  <dbl> <dbl>  <dbl> <dbl> <dbl>
#> 1 all      0.144 0.444 -0.339 0.620 -0.132 0.352  1000
#> 2 middle   0.130 0.411 -0.387 0.613 -0.108 0.316   331
#> 3 old      0.204 0.506 -0.304 0.662 -0.110 0.418   323
#> 4 young    0.101 0.375 -0.310 0.567 -0.139 0.287   346

## Example 2 ----
# Setup up specifications with specific additions to the formula
specs2 <- setup(data = example_data,
                y = c("y1", "y2"),
                x = c("x1", "x2"),
                model = "lm",
                distinct(example_data, group1),
                distinct(example_data, group2),
                controls = "c1",
                add_to_formula = "c2") # Always control for `c2`

# Run analysis using parallelization (2 cores and `future_pmap()`).
# Because "lm" is a base function, no further arguments necessary
results2 <- specr(specs2, workers = 2)
#> Models fitted based on 96 specifications
#> Cores used: 2 
#> 5.912 sec elapsed

# Summary of results
summary(results2, digits = 3)
#> Results of the specification curve analysis
#> -------------------
#> Technical details:
#> 
#>   Class:                          specr.object -- version: 0.3.0 
#>   Cores used:                     2 
#>   Duration of fitting process:    5.912 sec elapsed 
#>   Number of specifications:       96 
#> 
#> Descriptive summary of the specification curve:
#> 
#>  median   mad    min   max    q25  q75
#>   0.137 0.394 -0.292 0.675 -0.042 0.44
#> 
#> Descriptive summary of sample sizes: 
#> 
#>  median min  max
#>     250 160 1000
#> 
#> Head of the specification results (first 6 rows): 
#> 
#> # A tibble: 6 × 26
#>   x     y     model controls subsets    group1 group2 formula estimate std.error
#>   <chr> <chr> <chr> <chr>    <chr>      <fct>  <fct>  <glue>     <dbl>     <dbl>
#> 1 x1    y1    lm    1        middle & … middle female y1 ~ x…    0.596     0.101
#> 2 x1    y1    lm    1        old & fem… old    female y1 ~ x…    0.643     0.083
#> 3 x1    y1    lm    1        young & f… young  female y1 ~ x…    0.48      0.082
#> 4 x1    y1    lm    1        female     NA     female y1 ~ x…    0.602     0.054
#> 5 x1    y1    lm    1        middle & … middle male   y1 ~ x…    0.578     0.093
#> 6 x1    y1    lm    1        old & male old    male   y1 ~ x…    0.675     0.1  
#> # … with 16 more variables: statistic <dbl>, p.value <dbl>, conf.low <dbl>,
#> #   conf.high <dbl>, fit_r.squared <dbl>, fit_adj.r.squared <dbl>,
#> #   fit_sigma <dbl>, fit_statistic <dbl>, fit_p.value <dbl>, fit_df <dbl>,
#> #   fit_logLik <dbl>, fit_AIC <dbl>, fit_BIC <dbl>, fit_deviance <dbl>,
#> #   fit_df.residual <dbl>, fit_nobs <dbl>

## Example 3 ----
# Create custom function
glm2 <- function(formula, data) {
  glm(formula, data, family = gaussian())
}

# Setup specifications
specs3 <- setup(data = example_data,
                y = c("y1", "y2"),
                x = c("x1", "x2"),
                model = c("lm", "glm2"), # custom function added
                distinct(example_data, group1),
                distinct(example_data, group2),
                controls = "c1",
                add_to_formula = "c2")

# Because "glm2" is a custom function, we need to provide `furrr_options`
opts <- furrr_options(globals = list(glm2 = glm2))

# Run analysis using parallelization (4 cores and `future_pmap`).
results3 <- specr(specs3,
                  workers = 4,
                  .options = opts) # Options passed to to `future_pmap()`
#> Models fitted based on 192 specifications
#> Cores used: 4 
#> 10.975 sec elapsed

# Summary of results
summary(results3, digits = 3)
#> Results of the specification curve analysis
#> -------------------
#> Technical details:
#> 
#>   Class:                          specr.object -- version: 0.3.0 
#>   Cores used:                     4 
#>   Duration of fitting process:    10.975 sec elapsed 
#>   Number of specifications:       192 
#> 
#> Descriptive summary of the specification curve:
#> 
#>  median   mad    min   max    q25  q75
#>   0.137 0.394 -0.292 0.675 -0.042 0.44
#> 
#> Descriptive summary of sample sizes: 
#> 
#>  median min  max
#>     250 160 1000
#> 
#> Head of the specification results (first 6 rows): 
#> 
#> # A tibble: 6 × 28
#>   x     y     model controls subsets    group1 group2 formula estimate std.error
#>   <chr> <chr> <chr> <chr>    <chr>      <fct>  <fct>  <glue>     <dbl>     <dbl>
#> 1 x1    y1    lm    1        middle & … middle female y1 ~ x…    0.596     0.101
#> 2 x1    y1    lm    1        old & fem… old    female y1 ~ x…    0.643     0.083
#> 3 x1    y1    lm    1        young & f… young  female y1 ~ x…    0.48      0.082
#> 4 x1    y1    lm    1        female     NA     female y1 ~ x…    0.602     0.054
#> 5 x1    y1    lm    1        middle & … middle male   y1 ~ x…    0.578     0.093
#> 6 x1    y1    lm    1        old & male old    male   y1 ~ x…    0.675     0.1  
#> # … with 18 more variables: statistic <dbl>, p.value <dbl>, conf.low <dbl>,
#> #   conf.high <dbl>, fit_r.squared <dbl>, fit_adj.r.squared <dbl>,
#> #   fit_sigma <dbl>, fit_statistic <dbl>, fit_p.value <dbl>, fit_df <dbl>,
#> #   fit_logLik <dbl>, fit_AIC <dbl>, fit_BIC <dbl>, fit_deviance <dbl>,
#> #   fit_df.residual <dbl>, fit_nobs <dbl>, fit_null.deviance <dbl>,
#> #   fit_df.null <dbl>