Summarizing the Specifications Setup

summary method for class "specr.setup". Provides a short summary of the created specifications (the "multiverse") that lists all analytic choices, prints the function used to extract the parameters from the model. Finally, if print.specs = TRUE, it also shows the head of the actual specification grid.

# S3 method for specr.setup
summary(x, digits = 2, rows = 6, print.specs = TRUE)

Arguments

x

An object of class "specr.setup", usually, a result of a call to setup.

digits

The number of digits to use when printing the specification table.

rows

The number of rows of the specification tibble that should be printed.

print.specs

Logical value; if TRUE, a head of the specification tibble is returned and printed.

Value

A printed summary of an object of class specr.setup.

See also

Examples

# Setup specifications
specs <- setup(data = example_data,                   # data
               x = c("x1", "x2"),                     # independent variables
               y = c("y1", "y2"),                     # dependent variable
               model = c("lm", "glm"),                # model functions
               distinct(example_data, group3),        # subsets
               controls = c("c1", "c2", "c3"))        # covariates to be added
summary(specs)
#> Setup for the Specification Curve Analysis
#> -------------------------------------------
#> Class:                      specr.setup -- version: 0.3.0 
#> Number of specifications:   256 
#> 
#> Specifications:
#> 
#>   Independent variable:     x1, x2 
#>   Dependent variable:       y1, y2 
#>   Models:                   lm, glm 
#>   Covariates:               1 (none), c1, c2, c3, c1 + c2, c1 + c3, c2 + c3, c1 + c2 + c3 
#>   Subsets analyses:         C, B, A, all 
#> 
#> Function used to extract parameters:
#> 
#>   function (x) 
#> broom::tidy(x, conf.int = TRUE)
#> <environment: 0x7f7ee9e14398>
#> 
#> 
#> Head of specifications table (first 6 rows):
#> 
#> # A tibble: 6 × 7
#>   x     y     model controls subsets group3 formula     
#>   <chr> <chr> <chr> <chr>    <chr>   <fct>  <glue>      
#> 1 x1    y1    lm    1        C       C      y1 ~ x1 + 1 
#> 2 x1    y1    lm    1        B       B      y1 ~ x1 + 1 
#> 3 x1    y1    lm    1        A       A      y1 ~ x1 + 1 
#> 4 x1    y1    lm    1        all     NA     y1 ~ x1 + 1 
#> 5 x1    y1    lm    c1       C       C      y1 ~ x1 + c1
#> 6 x1    y1    lm    c1       B       B      y1 ~ x1 + c1
summary(specs, rows = 10)
#> Setup for the Specification Curve Analysis
#> -------------------------------------------
#> Class:                      specr.setup -- version: 0.3.0 
#> Number of specifications:   256 
#> 
#> Specifications:
#> 
#>   Independent variable:     x1, x2 
#>   Dependent variable:       y1, y2 
#>   Models:                   lm, glm 
#>   Covariates:               1 (none), c1, c2, c3, c1 + c2, c1 + c3, c2 + c3, c1 + c2 + c3 
#>   Subsets analyses:         C, B, A, all 
#> 
#> Function used to extract parameters:
#> 
#>   function (x) 
#> broom::tidy(x, conf.int = TRUE)
#> <environment: 0x7f7ee9e14398>
#> 
#> 
#> Head of specifications table (first 10 rows):
#> 
#> # A tibble: 10 × 7
#>    x     y     model controls subsets group3 formula     
#>    <chr> <chr> <chr> <chr>    <chr>   <fct>  <glue>      
#>  1 x1    y1    lm    1        C       C      y1 ~ x1 + 1 
#>  2 x1    y1    lm    1        B       B      y1 ~ x1 + 1 
#>  3 x1    y1    lm    1        A       A      y1 ~ x1 + 1 
#>  4 x1    y1    lm    1        all     NA     y1 ~ x1 + 1 
#>  5 x1    y1    lm    c1       C       C      y1 ~ x1 + c1
#>  6 x1    y1    lm    c1       B       B      y1 ~ x1 + c1
#>  7 x1    y1    lm    c1       A       A      y1 ~ x1 + c1
#>  8 x1    y1    lm    c1       all     NA     y1 ~ x1 + c1
#>  9 x1    y1    lm    c2       C       C      y1 ~ x1 + c2
#> 10 x1    y1    lm    c2       B       B      y1 ~ x1 + c2