Run specification curve analysis

This is the central function of the package. It can be used to run the specification curve analyses. It takes the data frame and vectors for analytical choices related to the dependent variable, the independent variable, the type of models that should be estimated, the set of covariates that should be included (none, each individually, and all together), as well as a named list of potential subsets. It returns a "result frame" which includes relevant statistics for each model as well as the analytical choices as factorial variables.

run_specs(df, y, x, model, controls = NULL, subsets = NULL,
  conf.level = 0.95)

Arguments

df
a data frame that includes all relevant variables

y
a vector of the dependent variables

x
a vevtor of the dependent variables

model
a vector of the type of models that should be estimated.

controls
a vector of the control variables that should be included. Defaults to none.

subsets
a list that includes named vectors

conf.level
the confidence level to use for the confidence interval. Must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.

Examples

# run specification curve analysis
results <- run_specs(df = example_data,
                     y = c("y1", "y2"),
                     x = c("x1", "x2"),
                     model = c("lm", "lm_gauss"),
                     controls = c("c1", "c2"),
                     subsets = list(group1 = unique(example_data$group1),
                                    group2 = unique(example_data$group2)))

#> Error in lm_gauss(data = structure(list(x1 = c(1.53391339253513, 1.22394130773124, 1.90713434712958, 1.69002857326579, 3.03701749905113, 3.87317404009697, 3.22399542495333, 3.96640841416119, 0.948535642178887, 1.60134178711307, 3.39583933449139, 2.96547494456049, 2.22138693578754, 1.12157680025732, -1.02227068122169, 2.96343478132845, 4.70259362847554, 2.6141152207728, 1.34050178297542, 3.44485153818207, 1.25785664126127, 1.3345923330968, 2.90407070833283, 3.15571768447671, 4.15194036800627, 2.27750053350017, 2.00625349839386, -1.18461966844781, 3.10557518427658, 1.9625808507495, 2.59996349064868, 2.59606479061518, 3.36975394010901, 1.02991583724548, 3.83397679590515, 3.27610525931375, 1.56840648420431, 2.69293916814712, 2.77869330199194, 2.93592496478434, 1.72651978144085, 2.07748590968902, 2.68446193435024, 1.05648539943912, 1.25675240514542, 3.87703619174841, 3.20661542919239, 3.80777228292253, 2.62874959989893, 1.76296424420807, 2.80988488778945, 2.54816691433773, 3.1874366460689, -0.490641870870351, 1.65734009068691, 1.45509237942399, 4.59629484390061, 2.71061281647445, 1.62272475137633, 1.90085934064485, 3.74612295589219, 1.87016525047249, 1.8618880605716, 3.12920634599063, 2.90732119623033, 0.115974000499349, 2.98293030291431, 2.34717450936098, 2.44006788969709, 0.63206071163034, 2.40104003858611, 0.869727578694577, 3.98138221894332, 2.60546565552088, 3.6029479481961, 2.20355290483291, 2.0773275212337, 2.32458094041894, 2.25342975500579, 2.14418993842559, 3.3643718903902, 3.9134184032876, 1.96856604633162, 1.94477669886165, 2.91339212848586, 1.7447428674305, 3.11308480991431, 1.5242538058584, 3.48325736143287, 2.04093242993148, 2.65640774287082, 2.48280160589486, 1.25862266805802, 1.72858876834932, 2.16722684913644, 2.49535536493328, 1.96669062180217, 1.28917886057613, 3.67469591252678, 2.3910103786485, 1.21919709290099, 3.64583263146528, 1.89581785643039, 1.71092813442421, 3.57765593212299, 3.11594581522161, 2.70159608656164, 3.20744182997903, 0.867892708439673, 3.44320405971443, 2.44199674573287, 1.86471835463218, 4.0365589055221, 3.87182403043371, 2.19075809666141, 2.22967294388324, 4.38284724313454, 3.23370188884346, 1.94713198012681, 1.38791483761627, 1.4800779777656, 2.05863639539502, 2.47450947054289, 3.26170007056903, 3.52500194302987, 4.21461155190099, 2.5893362850999, 2.18956410269417, 2.3545630799259, 4.50429650783292, 3.50918483092993, 3.92759446790688, 1.67204730931072, 3.36992857709054, 2.94057456181157, 3.26439125308828, 3.34850044937929, 1.73260349597302, 2.76525143331732, 1.99229743454776, 0.412757745259857, 1.29521882666987, 4.67104590334755, 2.81845463016069, 1.8815245258027, 2.04648317722987, 4.10078588946532, 2.15546491252551, 2.53795482981889, 1.90734812626636, 2.13715188052643, 2.92835003513492, 2.61787671975786, 2.99201446037805, 2.08363697392439, 0.849124089555124, 2.89959829145653, 0.891656837178819, 0.854864654813726, 2.52837972028242, 0.938578811972095, 3.06546168552806, -0.194433005319617, 4.66846313868724, 3.34237864975883, 3.87966819107731, 1.68771380511603, 4.5010127494246, 2.47865807815885, 3.30846119389307, 0.412032703622784, 3.52434983122512, 1.42518018474217, 3.00591519802114, 2.57375873139068, 1.62246720979004, 3.01741086112469, 0.475003878259969, 1.75782527688699, 2.77065186425249, 2.08129533148489, 3.3871006096992, 2.88290311648069, 0.881229830265771, 3.22994971217265, 2.54072018995303, 1.88093544912115, 2.6793390430831, 1.29262683953639, 2.66483560092587, 1.39101384122052, 1.25050330676116, 4.16906334996098, 2.80740426905371, 1.88727859198951, 2.0253668787392, 0.548365133627821, 4.13973290731384, 2.59835090907339, 2.8600083735503, 3.53610767267063, 1.31096123594059, 0.185616941837293, 1.92433685709914, 2.37322347363493, 2.96234590677905, 2.18654496107996, 4.15232393369548, 1.30907456482906, 2.98481202365195, 2.15648203054525, 0.267918299798451, 2.13341539098656, 3.87442798483711, 2.91716877290453, 3.73771731648257, 2.6185011466591, 1.03448377543768, 4.55149816116808, 1.64307693621333, 1.28775242667273, 2.87326254554788, 1.16941622087334, 3.25181359367564, 4.20211615484954, 3.36746948073028, 3.36868459887846, 3.5799177198534, 2.62214362523795, 1.83191226624105, 2.217263685973, 1.95485610433664, 2.27541877709057, 2.47969100837543, 3.83320542404769, 4.85831345029578, 2.86102133297485, 3.82767802796895, 1.96479914801332, 2.74170419667263, 2.39502762553916, 3.36485480868602, 4.29276560126371, 2.10796248393386, 2.62198374990789, 2.05783383167462, 1.50598649617438, 3.43121087988349, 1.31096495191501, 1.17275655015865), x2 = c(1.36971215879188, -0.238104407347796, 0.628281585136755, 0.540192437798822, -0.00477850171095184, 2.52160167991774, 0.949300587639871, 1.91759695022929, -1.57552333555491, -0.926949658866498, 1.27798993605891, 1.39013504816777, 0.779727979961508, -0.0832316458546501, -1.1169201590783, 3.40112173556891, 1.90263189050183, 2.44104141563339, 0.766094085383502, 0.1739901410451, -1.40737231244926, -0.86391534553445, 2.59093412722206, 1.7159971622693, 4.08586016195798, 0.625755792048509, 2.26309956592016, -1.21870843975786, 2.51161456977535, 0.471981368301565, 0.78721393560465, 0.864557553593621, 0.44083416787329, 1.79977371167975, 2.91531944596636, 2.6622224976667, 0.778281556491169, 3.05140177125899, 1.83125167052852, 2.8676598209239, 1.11823147598625, -1.36533536315023, 1.65059682240092, 1.01590529397931, 7.08502876313188e-05, 1.34590640646451, 1.64234092497158, 3.09885699360899, 3.30346791975004, 0.374877343993215, 2.90574573735142, 0.461049063511839, 1.23172281515574, -1.5264695787637, 1.18296665998606, 0.43149547414729, 2.53842833056502, 1.35931948772718, -0.255991217828574, 0.909182931652929, 1.69714065532075, 0.676422123202821, 1.542079587507, 1.64188267960085, 1.90048265250524, 0.676723334486714, 0.788747383688123, 1.32368615218625, 1.65483264525761, 0.879446299712365, 0.101040334565393, 0.559235615643488, 2.25375147302226, 2.93922597858855, 3.14967399394783, -1.50757048500591, -0.0118489457623361, 2.09106985892195, 1.2899608639547, 1.28951079739619, 0.0139551457012723, 1.89463799040836, 1.15773705972046, 3.14286535954312, 1.50330047960661, 0.140527296513724, 2.63810313040204, 2.0045852348025, 1.09427893967209, 0.233419728796916, 1.74928934699595, 0.392832958657479, 1.23803517044646, 1.03806936559775, 0.530974895224189, 0.523110740710364, 2.10521301489371, 0.695556312445009, 2.58332790765699, 0.685442336984946, 0.627921962078199, 1.90508979509488, 0.918242277653126, 2.2203456799077, 2.73156438637864, 0.346416884839539, 2.55636099109515, 0.536843882423192, -0.649823780813936, 2.25316645618431, 1.16220983770116, 0.0206146779458418, 0.828472656271068, 0.874143313056284, 0.417160658326917, 2.56405934389467, 3.62270976117128, 0.970002543585944, 0.257850774530833, 1.51287953655239, 0.643847526313869, 2.04054330150086, 1.97851694027794, 2.34138535783661, 1.51864721568008, 2.25866675784322, 1.21351014815462, 1.84880869998073, -0.150921684297074, 1.38734184759278, 2.11670220104557, 0.674614519958119, -0.619882968557955, 2.47650568113467, 2.30268158362967, 2.59839018865415, 1.29603075676817, 1.16459918922407, 1.48106975748245, 0.968894143298947, -0.156617833006485, 1.34517394436528, 1.00442664281039, 0.901294374782181, 0.28922193098881, 0.123488522444244, 2.90097031213957, 1.00868983035587, 3.67037092946932, 1.20274663508534, -0.565568257465065, 1.34267354926616, 3.04117926200459, 1.4488441962488, 3.07212242405597, 0.565287633952157, 0.46433701962981, -0.155068073365622, 3.75272225868688, 1.06691431861178, 1.13974302797662, 1.33524452977885, 0.969000527371868, 2.58440825248792, 1.58020174381495, 2.56653053657577, 1.06383353877926, 3.2091371081659, 0.307753565198541, 3.05345978003068, 0.667230877300853, 4.574060309223, 2.0294558021797, -0.0879619709083368, 0.50115098189522, 2.46934756624171, 1.34348134092091, -1.60371236751444, 1.47137963454302, -0.36034701817822, 1.29341969466535, 2.34997985259763, 1.69995153578001, 0.510378199091805, 3.15873058976486, 0.905064138043262, 3.0512950913773, -1.02798435339938, -0.403782449035207, 0.941144997684755, -0.653057162429069, 0.924258110223285, 3.80911445329899, 0.0179400403273147, 2.79353049029592, 2.16029969572183, -0.336367539460478, 3.80109961136295, 1.14723683805636, 2.75679116230208, 0.826838310790794, 2.29559310700595, -0.370073387928338, -1.45019310347859, 1.68466422173008, 0.0370711778310597, 1.7136592095613, 2.2833222760747, 1.85901282903256, 3.50066107667654, -0.611163404178786, 0.869619484776385, -0.178126454271006, 0.169369003009741, 3.95776217158601, 3.4517740267342, -0.0912024391746757, 0.896118839661425, 3.11572246275385, -0.0543924265197048, 0.876289095175665, 1.6527060371622, 2.19352691590369, 2.58805476113989, 2.51923233698616, 2.70331750315102, -0.643876304429945, 3.20773386208342, 1.1853400577738, 0.0764577110576156, 1.26642942366425, 1.3107710687945, 0.128842632192332, 1.56696828637096, 2.1776522507262, 2.6146702319927, 0.63204906343618, 3.14865596445009, -0.0829852770348301, 1.51114850042299, 1.47874063808592, 0.957686360280987, 2.54449576523078, 0.484964362434863, -0.674541213925125, 1.71825073705246, 1.52880365477945, 2.72887601938624, 0.633444292070411, 0.317987125282316), c1 = c(0.542490214532997, 0.140589065472275, 3.10025184924827, -0.156972297851147, 4.70940117260557, 1.96666371386208, 2.7322887275782, 3.20685186004695, 1.3439542335392, 1.91134672020215, 2.49943745958337, 5.26710027165904, 1.42888087994258, -0.0782202249868309, 2.55792895965824, 2.20427043430724, 3.04657086187497, 1.06763198507937, 4.20656035547328, 1.85426927149376, 2.78878916477846, 2.32023804708467, 1.46737449423091, 2.14754943799346, 1.26363473027641, -0.969040825437874, 2.65666108893293, 2.06378112584695, 1.34394678269232, 0.863725184057142, 0.103775773199643, -0.349168655431077, 4.05981361248921, 0.394428208513564, 1.60693815093285, 5.30852044914651, 1.65317619512722, 3.77783001083532, 2.19030164938172, 2.11559853903521, 3.57594428578569, -0.214145468635465, 1.42098897052977, 5.11334059535076, -0.34651857057947, 1.39805221483854, 3.52643802267967, 2.52775267807549, -1.53968011881452, 0.491780401907781, 1.76522860196125, 4.03851394031268, 0.779750875689732, 1.73366675619162, 4.13841385351925, 1.98720263902941, 3.86207958595578, 2.31384142921758, 0.21772148614942, 3.5210464365168, 0.718224519006378, 2.13393612308346, 0.29291909802744, 1.79212887181228, 6.15136720729653, 3.75819255977063, -1.49024583916698, -1.31356414487754, 1.43041340728603, 1.72566156169581, 7.99434406744349, 3.51742026172759, 3.82513303238926, 5.44869465783403, 3.77984574276078, 2.23615828012141, 4.35275031991069, 0.56662342420924, 1.0965497815972, 0.268919599906806, 1.51065481896494, 3.56508148622999, 1.15177105883908, 5.22593323459308, 0.937602498491563, 3.61618305719253, 3.86251532421379, 5.79117182511711, -0.439502349287577, -0.0536010753827418, 2.00438251500831, 1.71135247450271, -1.18349278810208, -0.209295411723517, 4.58982314179087, -2.0606080269375, 4.2815631894743, 4.45177898304371, 2.9788685317122, 2.71813512777769, 1.14342244380801, -4.09347389199759, 1.6840128137392, 0.657673264155148, 4.7440018291432, 4.3401934280584, 2.93529387104376, 1.23366478136321, 0.984871544473735, 1.25656210875324, 3.12894889675059, 2.75793552776246, 4.12577727372254, 4.13820749613048, 4.43614756247943, 2.13132819454586, 3.57471445086917, 3.0244109601229, 6.40765619803267, 1.82728811970084, 0.595511010060574, 2.38681315039406, 2.11269793509071, 5.27297709181763, 7.59770467587537, 2.60407818114397, 5.84524473658645, 1.38301862948344, 6.24013584951895, 3.02314745378762, 3.65309173679034, 3.18436638661971, 4.49530667411065, -0.281987592834586, 1.28903743091747, 5.12484076959798, 3.64278059827912, 2.05391758987772, 2.07965480749704, 4.80642455802796, 3.54909905580934, 5.04671590255601, -0.73195662703935, 0.501992570478598, 3.25543906755383, 2.12408424375167, 2.01347000614712, 3.71656868234882, -0.228928580242731, -0.881772624334171, 6.16396995819639, 4.64318942909539, 3.64652105811961, 1.89083062298151, -2.15951986755745, 4.17849082010188, 3.4098364216254, 2.73167031378627, 1.84801432297994, 1.68679646814729, -0.157527160781493, 0.5809462781858, 4.26631749970253, 4.20836309062767, 4.08757441081721, 1.89373530399946, 2.13681566673919, 1.72826548553917, 3.32248016011094, 4.11737582279427, 4.53163883512807, 1.02025493205861, 2.39551770094207, 2.74884224119605, 1.81348716859157, 2.86914838024078, 1.63007134098573, 3.79132000550343, 3.50373318158582, 0.996763021411182, 1.99901679468912, -0.311593270625557, 1.32380438077088, 3.16858543486057, -0.266315926828325, 2.90492023096826, 3.01319489177368, 1.64758006273004, 3.00693508878117, 2.30948116852625, 0.584476621028057, 1.86988668146539, 5.15361120685805, 1.2548380573694, 3.42903869173738, 2.38427178986885, 3.76103805615193, 2.97842262015345, 0.292945794802708, -0.0293941813876035, 2.70279350389298, 2.66418558285019, -0.112618673991069, 2.5715545380856, 4.20335461422297, 7.12762561937771, 0.816868373159786, 6.50703194124593, -1.33888693146688, 3.09588160315137, 4.60755740241534, 1.30954699268036, 3.37357995401211, 3.89501377938172, 0.930572569756093, 0.398027748048601, 3.6852791839986, 2.17337915817647, 4.72108074090162, 4.77129475228228, 3.7770171568566, 4.15944381553154, 4.00384040533905, 5.26871821905935, 1.80875751576318, 2.31813077940788, 3.38129949247426, 5.34640359792363, 1.70677571880833, 1.61006550653997, 1.03656610619575, -3.40455519227707, 6.44260945352773, 6.14327331273657, 2.38383127402963, 0.964012537151078, 3.06278457790613, -0.695940633816161, 1.99890686025204, -0.797321259841642, 2.89293036747878, 0.902948611391989, 1.7434098524141, 1.02213272551279, 1.55433930832346, 4.31280814444783, 0.0415076393917007, 0.197795578980887, 4.01924493000009, 4.09760596390637), c2 = c(3.89244345124217, 0.891195949001323, 5.58405741764772, 0.954775997198556, 2.4578714547592, 2.09237714829768, 3.36138819806135, -2.22701313204965, 0.711156193083337, 0.591221325073357, 4.74866818806734, 4.31646059635898, -3.66889521892431, 1.77526375759389, 1.82915319313252, 3.74758784336224, 4.08298928631226, 1.48802408406667, 1.22280940284654, 3.44419679028032, -0.450691836084937, 1.4203295972162, 3.01435440087567, 2.00008464773522, 3.94677451515945, 5.24689777251782, 2.49782860638495, 0.0928398916403022, 2.14264939979602, 0.32058867773284, 2.45890259658112, 4.25682729161988, 0.0831089046817461, 2.19562934438018, 0.215910500900439, -0.524841145883894, 4.37757793713531, -1.63223750454066, 4.19739358329951, 4.13029434248799, -1.30516115891895, 1.91073361541662, 3.62827966478296, 5.01873796491348, -0.526049738343034, 2.17271015214976, 1.29503968391579, 3.63142482243004, 3.82960780773163, 6.45033773144329, 2.72086163156582, 3.99831287241555, 1.56973669994695, 2.5427648289383, 4.34544171451611, 6.3727276918706, 4.51340621210318, 2.61005746553808, 4.27893900183212, 1.2693046613851, 2.3904560843267, 3.51911789966606, 2.66577965074891, 1.6771392628054, 8.05559906086734, 4.68306803657279, 2.13228173703736, 0.662586583049271, 2.82793030831017, -0.516003272290454, 1.70829902076594, 1.08250229785174, 3.11387932041281, 0.675386563041151, 3.62134916941477, 0.862831024979004, 1.36000025690564, 2.328525563657, 4.8294142670249, 1.60990267843476, 4.14996572286499, 2.80727932009096, 1.73366667700896, 2.18616642031054, 2.16414713358331, 0.211422071854074, 5.55490849732712, 4.51709431757484, 3.08422575349656, 4.21604146484179, 3.39714809167876, 2.01105314338055, 2.90257460620808, 4.15927218022112, 3.63330454575334, 1.53931655288491, 0.130456591745302, 2.99288753160912, 1.9305703620548, -0.169402688852414, 1.14791088746432, 1.28584598395168, 4.1161528193928, -0.55965074077394, 2.28958923755762, 0.382064243435751, 2.71098196762169, -0.0925546450900572, 4.88780089481114, 3.23525163614929, -0.94565365665929, 2.49474370760851, 5.19486110657979, 2.93023931598511, 5.22676166368708, 3.54719409406082, 6.44025232150829, 0.524387971106833, 4.52038967637929, 5.12017739506925, 1.17679452787229, 1.76423780200578, 3.47025470075987, -0.330100446794078, 6.18005663800548, 2.96311502216866, 1.07818020996183, 4.42246883035215, 3.44944003788187, 2.24444696027106, 4.04522401604983, 5.13908705121016, 3.08024419637285, 1.00206961453509, 1.02502113545668, 2.94739915598459, 2.49138827196181, 3.7929886003887, 3.93998320316972, 3.36266444437865, 3.20559248396575, 2.85484020420715, 0.137030837458604, 0.899284190170413, 4.21173080851073, 4.41691129448353, 0.837691496212698, -0.101528966783383, 2.27244232638924, 1.22435366024392, 4.26523522401293, 4.75638401657784, 2.80622678486426, -2.55442443362422, 3.97620299030224, 4.81600016584492, 2.14404828979818, 2.00142786354329, 6.40161172348523, 1.93650384041976, 5.00349331615979, 2.41900675496989, 3.02905314731224, 5.21248680323366, 4.25370804525185, 0.625636233885669, 6.38074969347023, 5.91486866477195, 3.93622432792852, 5.23165405582999, 3.72138323907563, 2.78261844702028, 3.31860664831364, 0.579327057851545, -0.372613143138413, 3.43238524645662, 1.87413252759974, 1.79084143100792, -0.211938921813293, 2.52981277514307, 5.0251411102937, 1.76014857758472, 0.562247537818574, -1.46125446581908, 2.79372724190542, 5.221880229604, 3.30014412000367, -1.32111994409994, 4.3463850259678, 0.356645968172647, 1.00366290054687, 5.82220004496965, 5.09930019467282, 2.26675302944542, 5.38567613407804, 4.55763909129522, 4.66148223068058, 3.11238303937523, 2.04983473200903, 2.18617391504294, 1.58097372810718, 4.58826310790098, -0.96425146342088, 3.54740657633181, 6.57069955257646, 3.20531112038827, 3.84648084495575, 3.02687996530323, 0.736921334952512, 1.70766438824374, 0.866881909965434, 1.05235100174768, 5.0600813436336, 4.98080280630954, 2.06031545873853, 2.19024191966736, 3.09630520425867, 3.97613385684577, 3.17704818114576, -1.60281423462487, 6.78674057812996, 1.05417801279644, 3.46355080489079, 6.41929567761267, 1.13362525489173, 6.35843560079058, 2.06925714918096, 0.73641611634634, 3.3518150891444, 0.551668804940797, 2.37934391524645, 1.1806092677673, 3.09247137590688, 3.06759880497182, 3.80444411591079, 3.80134026618799, 2.00467213957233, 4.02191360161614, 3.65291616070631, 2.41177535327734, 1.7117742751055, 1.41206464201505, 4.5383020381332, 5.37322197465337, 3.17340221089288, 2.28924874299311, 1.91959755290348, 2.80806654653273, -0.0415306892698388, 1.20858909564575), y1 = c(23.5005426129189, -3.67827168936398, 32.7131061133845, -0.244454281227271, 19.9848674586699, 19.0431552850219, 21.0327670215479, 6.31180412256954, -2.41479521413472, -1.48899870070494, 30.4138926186555, 40.1039094802155, -12.5437871840385, -6.48387987155716, -1.86434973945009, 37.0612675363215, 41.8260524925276, 7.63360621821322, 13.1965735164195, 16.9769064608406, 0.689188562861538, -8.09961945573062, 25.2667875828873, 10.1072595775099, 41.4297053986807, 14.5804927015303, 32.0406712936604, 4.47429869629548, 17.8214484258659, 2.52397397643881, 0.345729844631728, 11.7488746342076, 16.3662451456903, 15.210478375961, 8.61154972289411, 12.2378745536006, 14.9192102390833, -1.29451132424521, 32.2164655694916, 46.5423365450916, 5.51665054281679, -1.54159355211512, 23.4450914237213, 37.8668924421713, -4.9529251719713, 12.5378102454553, 21.2976910849858, 37.5338683078558, 31.6137496059653, 15.9483427885096, 28.3497474772393, 25.3202210071111, 7.13404536009426, 2.03975245335748, 29.0346602055074, 21.0689263355411, 52.4364667975185, 13.968858347971, 4.97601993661462, 15.9371173929117, 12.498052673131, 17.8310922100692, 11.8239207214737, 18.6600928494688, 67.782585080468, 22.141886109327, -2.56056243650291, -8.20499082385486, 20.8547118790797, 0.808992465468567, 30.9659155885615, 12.6015029814408, 38.9256856691924, 23.0900314506969, 40.0128727642033, -1.59828230461104, 14.9358239369075, 9.68068576309314, 26.6684935976424, 9.46268820608156, 14.1866749839326, 36.4906531308802, 10.0270974869494, 39.7068352940207, 13.1227831856521, 5.40140551022172, 62.1900310035506, 47.2828148750286, 18.3423556166749, 13.8476867913377, 24.0097745442788, 10.3308235778687, 12.2390620076288, 15.2984594656769, 24.0218390205423, 3.2637269818268, 16.6829535323414, 20.3229678972183, 21.5678000964287, 9.37652262013782, 5.35821165155344, -2.94199614640289, 19.325878158052, 8.30983291548135, 33.6334027611272, 16.5562902424411, 35.6445596083218, 2.93086223198897, 8.92058968792633, 24.6993798687005, 15.0583766513681, 0.123509345047529, 34.0455604816615, 29.1928814273576, 22.2174952782608, 37.1060599498936, 80.0859551005714, 10.5727690138009, 27.1037471815742, 31.3515430354835, 4.70236513292407, 21.2234427724982, 27.9472414736128, 14.759674632923, 58.6820098575812, 34.8543406476475, 26.8776765335596, 29.6919471885942, 30.0257764359071, 20.817888678441, 36.2591970138245, 35.7243712216993, 16.8724918685374, 5.6139640434874, 9.90371999047751, 37.5326035532751, 27.0278306543604, 21.8870988919901, 33.8876530610794, 28.2782008552871, 13.1421517732836, 22.4985283337444, -6.05758214677035, 2.82716727551209, 15.881604690834, 19.1606882338851, 17.1262825830188, 7.42640874550325, 17.7015138968977, 7.56930990990354, 26.4115326423124, 39.7002439268272, 29.3549934385143, -6.93325098740308, 21.4175631155861, 19.7417985734399, 11.9714945113889, 3.26246674625645, 69.8229700130116, 6.95157190719703, 17.8001792974198, 19.0006912039353, 24.1447054743173, 57.9549539644724, 42.4268277331405, 16.5223831627893, 31.9256188615293, 66.3762861442398, 17.1730293353475, 58.1949555186218, 17.8156785371303, 32.248322827683, 18.206441992711, 10.3001595285866, -1.60024164789795, 30.8532953889979, 10.3348884356328, 0.586051214189981, 5.97650448385907, 1.49046400542667, 27.8730989685414, 13.5889494074751, 6.23096544164539, 11.3876787313457, 23.6764897673659, 34.4805954295249, 40.6333746005458, 2.80617526121911, 12.8087323326101, 3.71157653918062, 3.77122957987845, 28.3783949743726, 66.7462764397042, 14.4537437996217, 52.3770642818107, 36.2387912287553, 8.23487502924275, 46.6644926861633, 4.22647630206687, 14.2726216644345, 17.71969322984, 30.3996155172684, -11.6155261928978, -0.742714917845256, 58.3151854336312, 31.5027876547825, 23.2257416677449, 40.5196189084967, -1.73034902110409, 24.5822001106221, 8.88665926510392, 6.75546388108085, 10.4516529908519, 24.3503969702755, 27.8442242795824, 25.0539667128771, 13.3036843439277, 14.3376588556843, 53.1587128205802, 4.99168270247313, 40.5029633956699, 21.7209460289929, 39.8699119480955, 62.4692209110998, 20.0285231489137, 57.9086129857734, 3.93256096125936, 29.5144053957539, 15.6909737732224, 7.37931031628266, 14.8946578842018, -2.67283168467732, 30.2407720928838, 36.2994856050627, 35.066392854894, 35.9239654029065, 14.8056242392391, 27.3626859543171, 17.7660845563624, 14.2382082110251, 15.6534784422608, 8.25484991838261, 43.5597729310199, 15.0374435405618, 1.89447760557799, 30.7903073825704, 13.7498465499554, 24.897530460327, 7.72341438444326, 7.27534782817172), y2 = c(10.4269278010497, 4.23031895523512, 20.5251919545908, 3.95461497955788, 14.1638082475462, 4.36014733470599, 10.6594029166602, 8.2002746828304, -6.0171519984009, -2.36662193381076, 23.9773236102184, 28.8871656742099, 3.04498039760456, -3.34946169968207, 8.84572558764311, 15.2374526677119, 12.6855454405332, -2.61026225384417, 18.4589069383415, 1.83472795352632, 1.13032095399871, 4.23191871056429, 17.7447947286866, 6.68786375985335, 7.61233169818869, 4.73321016455046, 14.4583257948467, 6.13224990798151, 3.76758406581309, 6.67540805743168, -0.849314238281894, -3.69237031041778, -4.53887900160228, 13.7338519113839, -3.28809981448017, -1.50792546103621, 10.8896722983187, 10.3149940181823, 23.6170089147653, 10.7669113789664, 6.64583848968661, -5.6006008887666, 15.6004017213678, 19.6943049991412, -4.16326839256625, 0.359661617317854, 17.4931219463201, 16.2527241999831, 5.91179241688438, 24.2133944061282, 4.79789369269401, 14.7876889610489, 1.70969045712748, 7.28574361989699, 17.890942500664, 10.6423825769116, 20.712326249644, 6.05407210946338, -1.11783844814306, 17.3876280440167, -2.21304021702759, 10.7381265667256, 3.82644526193369, 12.1875344153883, 46.3406043218193, 19.7333889512752, -13.6688615731725, -6.22312275610521, 15.764814197912, 2.78270468172, 32.6359919687363, 16.7230833900364, 19.9448596171245, 17.1499705056286, 12.8485173275453, 3.03996748996541, 6.91803571652382, 3.49586572221312, 21.3218108184765, -0.235128312807641, 4.42663216377062, 19.3396228299832, 11.6379019634787, 19.4384795487353, 1.14867450474997, 2.32196766638867, 17.1148656242377, 13.8174216550618, 11.1655428863977, 9.95400102510855, 9.90222257670569, 3.8930231471518, 3.08554460385849, 10.7237739987432, 24.6401506833361, -12.0484565707659, 5.1115585457982, 15.1183897903075, 0.290497332408477, 4.65746526408317, -2.75289649074232, -14.9825000509521, 9.70920412428092, 2.4278400221872, 27.04841083624, 16.2250259707397, 16.8664526891252, -2.88739854836457, 10.5959716924459, 12.0498502991283, 12.5055311578448, 7.42957692214232, 33.1954513660951, 17.9399761168501, 10.8019048209946, 19.9898298760257, 32.0131069624442, 8.77303054234912, 21.0740607914112, 13.2735152743521, -1.61049203834022, 10.7830542037695, 20.2548843462083, 19.909740348935, 33.3099181604556, 14.882313546161, 20.1283429467968, 17.7176570658812, 20.372087013238, 3.96551708781856, 13.2073122692282, 24.1451848346367, 6.51481665197337, -17.4093065061144, 0.0993962121184233, 19.80618101693, 23.8396222074211, 16.5625576807425, 14.0334572241938, 23.0531151916996, 12.999143491031, 16.5372611416207, -22.1480843315324, 9.36368698519232, 14.2341229987403, 15.3670543102843, 5.26661064178017, 7.29282812266768, 0.979559755823922, 0.619789324029281, 36.5037707407544, 35.1580589237094, 22.2207718321013, -2.22261409337206, -3.36768821507112, 20.625813797704, 24.0030606926934, 5.88977473573907, 19.0133187661152, 0.366592270625857, 5.60871026881474, 1.67914307170651, 20.8216432912991, 24.5154261239551, 10.3825171342113, 1.01270407549711, 16.9935891130771, 27.4770015008654, 21.6481384718887, 14.7661496193851, 21.1042099099023, 15.8138264290408, 20.7886478853631, 0.693774640711056, 0.712962779080952, 17.0192913182354, -2.78847691107079, 5.11693731923405, 2.95380642799602, -5.97297286727455, 15.2940746837812, -4.52901797638474, 7.22052733100825, -0.246174455417211, 5.87713595908097, 27.3954734264996, 18.9462419285807, -17.2181172839657, 21.0874636196618, 1.98515936440655, -5.60966701381343, 16.8856049234381, 28.3867953573763, 13.174432280312, 18.9899902309594, 11.2637817971704, 17.0505036883491, 13.4039427751284, -4.35849586421828, 0.265639464913419, 16.4983702311405, 15.9135530201705, -4.42043667983725, 15.093971177564, 30.5743071645085, 17.7074975857136, 10.3961471233174, 17.5947998283214, -1.2255678791367, -0.103368175422458, 16.389693616775, 0.784648441762064, 10.9261964691003, 12.302838420002, 6.26288633541946, 2.68334278741914, 10.5270300209479, 3.63605270027091, 18.2258584054085, 6.82228618660402, 22.2917032657182, 30.2313867683242, 21.7594779749042, 27.1674661025272, 10.6146389326473, 22.2971603786642, -0.0299135981366021, 26.3903171382297, 7.22046584337899, -1.7842684105025, 4.16258356824658, -4.3079420775276, 23.3293496925687, 24.9275849511498, 2.04582664110206, 1.24523245616324, -0.532862813125214, -6.82114074993268, 15.6805793150013, -9.40426880185912, 14.8882862671041, 15.7173757884951, 15.6829800261735, 9.85783137564152, 4.2181891724736, 16.5597533145859, 14.5618186661505, -3.92121172848557, 15.4188954381137, 14.7664713758045), group1 = c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0), group2 = structure(c(1L, 2L, 1L, 1L, 3L, 2L, 2L, 1L, 3L, 3L, 1L, 1L, 1L, 3L, 2L, 2L, 2L, 3L, 2L, 3L, 2L, 3L, 2L, 3L, 2L, 2L, 1L, 1L, 3L, 1L, 3L, 3L, 3L, 1L, 3L, 3L, 3L, 2L, 1L, 2L, 3L, 3L, 1L, 1L, 3L, 3L, 1L, 1L, 2L, 2L, 3L, 2L, 2L, 3L, 3L, 3L, 2L, 3L, 3L, 2L, 3L, 3L, 2L, 1L, 2L, 3L, 3L, 2L, 1L, 1L, 1L, 2L, 2L, 2L, 3L, 3L, 1L, 2L, 1L, 2L, 2L, 2L, 1L, 2L, 2L, 3L, 3L, 3L, 1L, 1L, 2L, 2L, 1L, 3L, 2L, 3L, 3L, 2L, 3L, 1L, 2L, 2L, 1L, 1L, 2L, 1L, 1L, 3L, 1L, 2L, 2L, 3L, 1L, 2L, 3L, 1L, 2L, 2L, 3L, 2L, 1L, 2L, 1L, 1L, 3L, 2L, 3L, 2L, 1L, 3L, 3L, 1L, 2L, 3L, 3L, 3L, 1L, 1L, 1L, 2L, 3L, 3L, 3L, 1L, 3L, 1L, 2L, 3L, 2L, 1L, 1L, 1L, 3L, 3L, 3L, 1L, 1L, 2L, 1L, 3L, 3L, 2L, 1L, 2L, 3L, 3L, 3L, 1L, 1L, 3L, 3L, 1L, 1L, 3L, 3L, 3L, 3L, 3L, 3L, 3L, 3L, 3L, 1L, 3L, 2L, 1L, 2L, 3L, 2L, 2L, 1L, 1L, 2L, 1L, 2L, 3L, 1L, 2L, 3L, 3L, 1L, 3L, 3L, 2L, 1L, 3L, 2L, 3L, 1L, 3L, 1L, 1L, 3L, 3L, 2L, 2L, 3L, 3L, 3L, 2L, 3L, 1L, 2L, 3L, 1L, 2L, 3L, 2L, 2L, 2L, 2L, 3L, 2L, 2L, 3L, 3L, 3L, 3L, 1L, 3L, 2L, 1L, 2L, 3L, 3L, 3L, 1L, 3L, 1L, 3L), .Label = c("A", "B", "C"), class = "factor"),     filter = c("group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0", "group1 = 0",     "group1 = 0")), class = "data.frame", row.names = c(NA, -250L)), formula = "y1 ~ x1 + c1 + c2"): could not find function "lm_gauss"

# Check results frame
results

#> Error in eval(expr, envir, enclos): object 'results' not found