Summarise specification curve results

This function allows to inspect results of the specification curves by returning a comparatively simple summary of the results. These results can be returned for specific analytical choices.

summarise_specs(
  df,
  var = estimate,
  stats = lst(median, mad, min, max, q25 = function(x) quantile(x, prob = 0.25), q75 =
    function(x) quantile(x, prob = 0.75)),
  group = NULL
)

Arguments

df
a data frame containing the choices and results of each specification (resulting from run_specs).

var
which variable should be evaluated? Defaults to estimate (the effect sizes computed by run_specs). No need to use closures.

stats
list object (lst()) of summary functions.

group
a grouping factor (e.g., "subsets"). Several grouping variables can be passed. Defaults to NULL.

Value

Examples

# Run specification curve analysis
results <- run_specs(df = example_data,
                     y = c("y1", "y2"),
                     x = c("x1", "x2"),
                     model = c("lm"),
                     controls = c("c1", "c2"),
                     subsets = list(group1 = unique(example_data$group1),
                                    group2 = unique(example_data$group2)))

# Basic example
summarise_specs(results)

#> # A tibble: 1 x 7
#>   median   mad   min   max   q25   q75   obs
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1   3.59  4.56 -2.05  9.58  1.03  7.63   123

# Summary of specific analytical choices
print(summarise_specs(results, group = c("subsets", "x")), n = 24)

#> # A tibble: 24 x 9
#> # Groups:   subsets [12]
#>    subsets                 x     median   mad     min   max     q25   q75   obs
#>    <chr>                   <chr>  <dbl> <dbl>   <dbl> <dbl>   <dbl> <dbl> <dbl>
#>  1 all                     x1      3.34 4.71  -0.0545  6.92  0.283   6.50   500
#>  2 all                     x2      4.48 4.66   1.09    8.14  1.34    7.67   500
#>  3 group1 = 0              x1      2.65 3.90  -0.227   5.85  0.0336  5.24   250
#>  4 group1 = 0              x2      4.05 4.33   0.653   8.07  1.29    7.08   250
#>  5 group1 = 0 & group2 = A x1      3.30 0.716  2.26    4.24  2.80    3.64    72
#>  6 group1 = 0 & group2 = A x2      4.06 4.24   0.482   8.20  1.52    6.88    72
#>  7 group1 = 0 & group2 = B x1      3.86 4.63   0.0819  8.96  1.06    7.10    76
#>  8 group1 = 0 & group2 = B x2      4.21 4.50   0.603   8.75  1.33    7.40    76
#>  9 group1 = 0 & group2 = C x1      1.41 4.72  -1.92    4.92 -1.76    4.61   102
#> 10 group1 = 0 & group2 = C x2      3.72 4.58   0.299   7.57  0.753   6.58   102
#> 11 group1 = 1              x1      4.20 5.59   0.153   8.27  0.584   7.84   250
#> 12 group1 = 1              x2      4.91 5.09   1.03    9.03  1.74    8.26   250
#> 13 group1 = 1 & group2 = A x1      5.89 3.98   2.79    9.20  3.44    8.44    72
#> 14 group1 = 1 & group2 = A x2      5.14 4.27   2.05    8.81  2.34    8.05    72
#> 15 group1 = 1 & group2 = B x1      4.88 5.96   0.608   9.28  1.09    8.85    86
#> 16 group1 = 1 & group2 = B x2      5.48 4.78   1.19    9.58  2.38    8.38    86
#> 17 group1 = 1 & group2 = C x1      2.38 6.26  -2.05    6.86 -1.76    6.58    92
#> 18 group1 = 1 & group2 = C x2      4.19 5.93  -0.258   8.65  0.326   8.11    92
#> 19 group2 = A              x1      4.56 2.33   2.52    6.67  3.11    5.90   144
#> 20 group2 = A              x2      4.89 4.49   1.64    8.25  1.98    7.78   144
#> 21 group2 = B              x1      4.35 5.16   0.474   8.92  0.972   7.97   162
#> 22 group2 = B              x2      4.48 4.65   1.05    8.27  1.62    7.65   162
#> 23 group2 = C              x1      1.79 5.38  -1.99    5.71 -1.79    5.44   194
#> 24 group2 = C              x2      4.03 5.14   0.207   8.11  0.582   7.49   194

# Summare of other estimates
summarise_specs(results, var = "p.value", group = "controls")

#> # A tibble: 4 x 8
#>   controls             median          mad      min   max      q25     q75   obs
#>   <chr>                 <dbl>        <dbl>    <dbl> <dbl>    <dbl>   <dbl> <dbl>
#> 1 c1            0.0000287          4.25e-5 9.94e-47 0.921 1.55e- 8 0.00997   123
#> 2 c1 + c2       0.00000000671      9.94e-9 5.08e-89 0.862 3.67e-18 0.00219   123
#> 3 c2            0.0000275          4.07e-5 8.09e-63 0.994 1.28e-13 0.0817    123
#> 4 no covariates 0.000484           7.17e-4 8.62e-40 0.656 3.81e- 9 0.0597    123